spin out types, utilities
