+++ apps/client/app/_utilities/formatDateISO.ts
@@ -1,0 +1,3 @@
+export const formatDateISO = (date: Date) => {
+  return new Date(date).toISOString().split('T')[0];
+};
